Cards updated to component. Style of Project Card updated. Mongoose strictQuery set to false. Jest dependency added.
