+++ Mern Stack/backend/package.json
@@ -4,9 +4,22 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --coverage --coverageDirectory=output/coverage/jest",
     "start": "node server.js",
     "dev": "nodemon server.js"
+  },
+  "jest": {
+    "coverageReporters": [
+      "text"
+    ],
+    "reporters": [
+      "default",
+      "jest-junit"
+    ]
+  },
+  "jest-junit": {
+    "output": "output/coverage/junit/junit.xml",
+    "usePathForSuiteName": "true"
   },
   "keywords": [],
   "author": "",
@@ -16,6 +29,8 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "jest": "^29.5.0",
+    "jest-junit": "^15.0.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.10.0",
     "validator": "^13.9.0"
