+++ Mern Stack/frontend/src/components/projectCard.css
@@ -1,0 +1,33 @@
+.card-container {
+    width: 70%;
+    height: 600px;
+  }
+  
+  .cards {
+    position: absolute;
+    background-color: #fff;
+    width: 100%;
+    height: 600px;
+    border-radius: 20px;
+  }
+  
+  .card-title {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 25px;
+    margin: 10px;
+  }
+  
+  .card-body {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: 30px;
+    font-size: 20px;
+  }
+  
+  .card-tag {
+    font-size: 15px;
+  }
+
