+++ Mern Stack/frontend/src/components/search.css
@@ -1,6 +1,3 @@
-
-  
-
   .search-container {
     display: flex;
     align-items: center;
